refactor: remove unused Result utility functions
- Keep only production-used functions in app/utils/result.ts (ok, err, isErr)
- Move test-only utility functions to tests/test-utils.ts for test isolation
- Remove completely unused tryCatchAsync function
- Update exports in app/utils/index.ts to only expose production functions
- Update Result test file to import from test-utils

This reduces production bundle size by ~190 lines of unused code while
maintaining all existing production functionality and test coverage.
